requirements.txt
python-telegram-bot>=21.0
python-telegram-bot[job-queue]
aiohttp>=3.9
python-dotenv>=1.0
psutil>=5.9